провел рефакторинг объектов view, моделей, расширений и application

+++ Blogs/Blogs/Screens/Choice Color/ChoiceColorViewController.swift
@@ -138,6 +138,6 @@ extension ChoiceColorViewController: UITableViewDelegate, UITableViewDataSource 
     }
     
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return section.rows[indexPath.row].cellHeight
+        return CGFloat(section.rows[indexPath.row].cellHeight)
     }
 }
