+++ Blogs/Blogs/Screens/Choice Language/ChoiceLanguageViewController.swift
@@ -138,7 +138,7 @@ extension ChoiceLanguageViewController: UITableViewDelegate, UITableViewDataSour
     }
     
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return section.rows[indexPath.row].cellHeight
+        return CGFloat(section.rows[indexPath.row].cellHeight)
     }
 }
